% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{bundleResults}
\alias{bundleResults}
\title{Bundles the results in a zip file}
\usage{
bundleResults(outputFolder, databaseId)
}
\arguments{
\item{outputFolder}{Folder to store the results}

\item{databaseId}{ID of your database, this will be used as subfolder for the results.}
}
\description{
\code{bundleResults} creates a zip file with results in the outputFolder
}
