% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceChecks.R
\name{performanceChecks}
\alias{performanceChecks}
\title{The performane checks (for v5.x)}
\usage{
performanceChecks(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema,
  vocabDatabaseSchema = cdmDatabaseSchema,
  oracleTempSchema = resultsDatabaseSchema,
  sourceName = "",
  sqlOnly = FALSE,
  outputFolder = "output",
  verboseMode = TRUE
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that contains OMOP CDM schema.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_instance.dbo'.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{vocabDatabaseSchema}{String name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want all temporary tables to be managed. Requires create/insert permissions to this database.}

\item{sourceName}{String name of the data source name. If blank, CDM_SOURCE table will be queried to try to obtain this.}

\item{sqlOnly}{Boolean to determine if Achilles should be fully executed. TRUE = just generate SQL files, don't actually run, FALSE = run Achilles}

\item{outputFolder}{Path to store logs and SQL files}

\item{verboseMode}{Boolean to determine if the console will show all execution steps. Default = TRUE}
}
\value{
\preformatted{                            An object of type \code{achillesResults} containing details for connecting to the database containing the results
}
}
\description{
\code{PerformanceChecks} runs a list of performance checks as part of the CDM inspection procedure
}
\details{
\code{PerformanceChecks} runs a list of performance checks as part of the CDM inspection procedure
}
