% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdmInspection.R
\name{cdmInspection}
\alias{cdmInspection}
\title{The main CdmInspection analyses (for v5.x)}
\usage{
cdmInspection(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema,
  scratchDatabaseSchema = resultsDatabaseSchema,
  vocabDatabaseSchema = cdmDatabaseSchema,
  oracleTempSchema = resultsDatabaseSchema,
  databaseName = "",
  databaseId = "",
  databaseDescription = "",
  analysisIds = "",
  smallCellCount = 5,
  runVocabularyChecks = TRUE,
  runDataTablesChecks = TRUE,
  runPerformanceChecks = TRUE,
  runWebAPIChecks = TRUE,
  baseUrl,
  sqlOnly = FALSE,
  outputFolder = "output",
  verboseMode = TRUE
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that contains OMOP CDM schema.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_instance.dbo'.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{scratchDatabaseSchema}{Fully qualified name of database schema that we can write temporary tables to. Default is resultsDatabaseSchema.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_scratch.dbo'.}

\item{vocabDatabaseSchema}{String name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want all temporary tables to be managed. Requires create/insert permissions to this database.}

\item{databaseName}{String name of the database name. If blank, CDM_SOURCE table will be queried to try to obtain this.}

\item{databaseId}{ID of your database, this will be used as subfolder for the results.}

\item{databaseDescription}{Provide a short description of the database.}

\item{analysisIds}{Analyses to run}

\item{smallCellCount}{To avoid patient identifiability, cells with small counts (<= smallCellCount) are deleted. Set to NULL if you don't want any deletions.}

\item{runVocabularyChecks}{Boolean to determine if vocabulary checks need to be run. Default = TRUE}

\item{runDataTablesChecks}{Boolean to determine if table checks need to be run. Default = TRUE}

\item{runPerformanceChecks}{Boolean to determine if performance checks need to be run. Default = TRUE}

\item{runWebAPIChecks}{Boolean to determine if WebAPI checks need to be run. Default = TRUE}

\item{baseUrl}{WebAPI url, example: http://server.org:80/WebAPI}

\item{sqlOnly}{Boolean to determine if Achilles should be fully executed. TRUE = just generate SQL files, don't actually run, FALSE = run Achilles}

\item{outputFolder}{Path to store logs and SQL files}

\item{verboseMode}{Boolean to determine if the console will show all execution steps. Default = TRUE}
}
\value{
\preformatted{                            An object of type \code{achillesResults} containing details for connecting to the database containing the results
}
}
\description{
\code{CdmInspection} runs a list of checks as part of the CDM inspection procedure
}
\details{
\code{CdmInspection} runs a list of checks as part of the CDM inspection procedure
}
